requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
autogen-agentchat==0.4.7
autogen-core==0.4.7
autogen-ext==0.4.7
azure-ai-documentintelligence==1.0.0
azure-ai-inference==1.0.0b9
azure-core==1.32.0
azure-identity==1.19.0
beautifulsoup4==4.13.3
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cobble==0.1.4
colorama==0.4.6
cryptography==44.0.0
defusedxml==0.7.1
Deprecated==1.2.17
distro==1.9.0
et_xmlfile==2.0.0
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.5.0
isodate==0.7.2
jiter==0.8.2
jsonref==1.1.0
lxml==5.3.1
mammoth==1.9.0
markdownify==1.0.0
markitdown==0.0.1a4
msal==1.31.1
msal-extensions==1.2.0
multidict==6.1.0
numpy==2.2.3
olefile==0.47
openai==1.60.1
openpyxl==3.1.5
opentelemetry-api==1.29.0
pandas==2.2.3
pathvalidate==3.2.3
pdfminer.six==20240706
pillow==11.1.0
playwright==1.50.0
portalocker==2.10.1
propcache==0.3.0
protobuf==5.29.3
puremagic==1.28
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
pyee==12.1.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-pptx==1.0.2
pytz==2025.1
pywin32==308
regex==2024.11.6
requests==2.32.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SpeechRecognition==3.14.1
tiktoken==0.8.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wrapt==1.17.2
xlrd==2.0.1
XlsxWriter==3.2.2
yarl==1.18.3
youtube-transcript-api==0.6.3
zipp==3.21.0
